+++ src/objectwizards/wizards/alEnumExtWizard.ts
@@ -1,7 +1,6 @@
 'use strict';
 
 import { ALObjectWizard } from "./alObjectWizard";
-import { ALLangServerProxy } from '../../allanguage/alLangServerProxy';
 import { ALEnumExtWizardData } from "./alEnumExtWizardData";
 import { ALEnumExtWizardPage } from "./alEnumExtWizardPage";
 import { DevToolsExtensionContext } from "../../devToolsExtensionContext";
@@ -24,6 +23,7 @@ export class ALEnumExtWizard extends ALObjectWizard {
         let wizardData : ALEnumExtWizardData = new ALEnumExtWizardData();
         wizardData.objectId = objectId.toString();
         wizardData.objectName = '';
+        wizardData.firstValueId = this._toolsExtensionContext.alLangProxy.getIdRangeStart(settings.getDestDirectoryUri());
         let wizardPage : ALEnumExtWizardPage = new ALEnumExtWizardPage(this._toolsExtensionContext, settings, wizardData);
         wizardPage.show();
     }